feat: add blogs use-cases

+++ src/features/blogs/api/blogs.controller.ts
@@ -24,6 +24,14 @@ import { PostsWithPaginationViewModel } from '../../posts/api/models/output/post
 import { PostCreateModel } from '../../posts/api/models/input/post.input.model';
 import { BlogDbType } from '../../types';
 import { BasicAuthGuard } from '../../auth/guards/basic-auth.guard';
+import { CommandBus } from '@nestjs/cqrs';
+import { GetAllQueryBlogsCommand } from '../application/use-cases/getAllQueryBlogs.useCase';
+import { CreateBlogByAdminCommand } from '../application/use-cases/createBlogByAdmin.useCase';
+import { GetBlogByIdCommand } from '../application/use-cases/getBlogById.useCase';
+import { GetQueryPostsCommand } from '../../posts/application/use-cases/getQueryPosts.useCase';
+import { CreatePostByAdminWithBlogIdCommand } from '../../posts/application/use-cases/createPostByAdminWithBlogId.useCase';
+import { GetQueryBlogByIdCommand } from '../application/use-cases/getQueryBlogById.useCase';
+import { UpdateBlogByAdminCommand } from '../application/use-cases/updateBlogByAdmin.useCase';
 
 @Controller('blogs')
 export class BlogsController {
@@ -32,6 +40,7 @@ export class BlogsController {
     private readonly blogService: BlogsService,
     private readonly postQueryRepository: PostsQueryRepository,
     private readonly postsService: PostsService,
+    private commandBus: CommandBus,
   ) {}
 
   @Get()
@@ -46,12 +55,14 @@ export class BlogsController {
     },
   ) {
     const foundBlogs: BlogWithPaginationViewModel =
-      await this.blogQueryRepository.findAllBlogs(
-        query.searchNameTerm,
-        query.sortBy,
-        query.sortDirection,
-        query.pageNumber,
-        query.pageSize,
+      await this.commandBus.execute(
+        new GetAllQueryBlogsCommand(
+          query.searchNameTerm,
+          query.sortBy,
+          query.sortDirection,
+          query.pageNumber,
+          query.pageSize,
+        ),
       );
     return foundBlogs;
   }
@@ -59,7 +70,9 @@ export class BlogsController {
   @UseGuards(BasicAuthGuard)
   @Post()
   async createBlogByAdmin(@Body() inputBlogModel: BlogCreateModel) {
-    return this.blogService.createBlogByAdmin(inputBlogModel);
+    return this.commandBus.execute(
+      new CreateBlogByAdminCommand(inputBlogModel),
+    );
   }
 
   @Get(':blogId/posts')
@@ -74,16 +87,20 @@ export class BlogsController {
     },
     @Body() userId?: string,
   ) {
-    const foundBlogById = await this.blogService.findBlogById(blogId);
+    const foundBlogById = await this.commandBus.execute(
+      new GetBlogByIdCommand(blogId),
+    );
     if (foundBlogById) {
       const foundPostsByBlogId: PostsWithPaginationViewModel =
-        await this.postQueryRepository.findPosts(
-          query.pageNumber,
-          query.pageSize,
-          query.sortBy,
-          query.sortDirection,
-          userId,
-          blogId,
+        await this.commandBus.execute(
+          new GetQueryPostsCommand(
+            query.pageNumber,
+            query.pageSize,
+            query.sortBy,
+            query.sortDirection,
+            userId,
+            blogId,
+          ),
         );
       return foundPostsByBlogId;
     } else {
@@ -97,13 +114,16 @@ export class BlogsController {
     @Param('blogId') blogId: string,
     @Body() inputPostModel: PostCreateModel,
   ) {
-    const foundBlog: BlogDbType | null =
-      await this.blogService.findBlogById(blogId);
+    const foundBlog: BlogDbType | null = await this.commandBus.execute(
+      new GetBlogByIdCommand(blogId),
+    );
     if (foundBlog) {
-      return await this.postsService.createPostByAdminWithBlogId(
-        inputPostModel,
-        foundBlog._id,
-        foundBlog.name,
+      return await this.commandBus.execute(
+        new CreatePostByAdminWithBlogIdCommand(
+          inputPostModel,
+          foundBlog._id,
+          foundBlog.name,
+        ),
       );
     } else {
       throw new NotFoundException();
@@ -112,7 +132,7 @@ export class BlogsController {
 
   @Get(':id')
   async findBlogById(@Param('id') blogId: string) {
-    return await this.blogQueryRepository.findBlogById(blogId);
+    return await this.commandBus.execute(new GetQueryBlogByIdCommand(blogId));
   }
 
   @UseGuards(BasicAuthGuard)
@@ -122,9 +142,13 @@ export class BlogsController {
     @Param('id') blogId: string,
     @Body() inputBlogModel: BlogUpdateModel,
   ) {
-    const foundBlog = await this.blogService.findBlogById(blogId);
+    const foundBlog = await this.commandBus.execute(
+      new GetBlogByIdCommand(blogId),
+    );
     if (foundBlog) {
-      await this.blogService.updateBlogByAdmin(blogId, inputBlogModel);
+      await this.commandBus.execute(
+        new UpdateBlogByAdminCommand(blogId, inputBlogModel),
+      );
     } else {
       throw new NotFoundException();
     }
@@ -134,8 +158,15 @@ export class BlogsController {
   @Delete(':id')
   @HttpCode(204)
   async removeBlogByAdmin(@Param('id') blogId: string) {
-    const blogRemoved = await this.blogService.removeBlogByAdmin(blogId);
-    if (!blogRemoved) {
+    const foundBlog = await this.commandBus.execute(
+      new GetBlogByIdCommand(blogId),
+    );
+    if (foundBlog) {
+      const blogRemoved = await this.blogService.removeBlogByAdmin(blogId);
+      if (!blogRemoved) {
+        throw new NotFoundException();
+      }
+    } else {
       throw new NotFoundException();
     }
   }
